refactor: update service layer to handle custom exceptions

+++ src/main/java/com/gabrielgermano/manager/service/EmployeeService.java
@@ -1,5 +1,7 @@
 package com.gabrielgermano.manager.service;
 
+import com.gabrielgermano.manager.exception.EmailAlreadyExistsException;
+import com.gabrielgermano.manager.exception.EmployeeNotFoundException;
 import com.gabrielgermano.manager.model.Employee;
 import com.gabrielgermano.manager.repository.EmployeeRepository;
 import org.slf4j.Logger;
@@ -23,7 +25,7 @@ public class EmployeeService {
 
     public Employee getEmployeeById(Long id) {
         logger.info("Fetching employee by id: {}", id);
-        return employeeRepository.findById(id).orElseThrow();
+        return employeeRepository.findById(id).orElseThrow(() -> new EmployeeNotFoundException(id));
     }
 
     public List<Employee> getAllEmployees() {
@@ -33,12 +35,18 @@ public class EmployeeService {
 
     public Employee createEmployee(Employee employee) {
         logger.info("Creating new employee: name={}, email={}, position={}", employee.getName(), employee.getEmail(), employee.getPosition());
+
+        if (employeeRepository.existsByEmail(employee.getEmail())) {
+            logger.warn("Attempted to create an employee with an existing email address: {}", employee.getEmail());
+            throw new EmailAlreadyExistsException(employee.getEmail());
+        }
+
         return employeeRepository.save(employee);
     }
 
     public Employee updateEmployee(Long id, Employee employee) {
         logger.info("Updating employee with ID: {}", id);
-        Employee foundEmployee = employeeRepository.findById(id).orElseThrow();
+        Employee foundEmployee = employeeRepository.findById(id).orElseThrow(() -> new EmployeeNotFoundException(id));
 
         foundEmployee.setName(employee.getName());
         foundEmployee.setSalary(employee.getSalary());
@@ -51,6 +59,10 @@ public class EmployeeService {
 
     public void deleteEmployee(Long id) {
         logger.info("Attempting to delete employee with ID: {}", id);
+        if (!employeeRepository.existsById(id)) {
+            logger.warn("Attempted to delete an non existing employee with ID: {}", id);
+            throw new EmployeeNotFoundException(id);
+        }
         employeeRepository.deleteById(id);
     }
